docs: explain foodtruck initialization
Food trucks are given 3 random location and time objects.
This will be considered the food truck's schedule and will be used
later in the code to animate the position of the food trucks.
Initial testing of the foodtrucks initialization and marker rendering
has been done in the initialize function. The code needs to be
refactored so that the rendering happens in the view model.