fix: Random food truck position
Random food trucks were showing up on the edge of the viewing area.
This caused a poor user experience. By identifying the smallest
dimension of the screen and constraining the marker render area
to that dimension then food truck markers will always display
completely or not at all.

! dang it - if the screen is in landscape - then this doesn't work!

Next steps. Identify smallest screen dimension, height of the markers,
anchor point of the markers.

There are two cases depending on the orientation if the goals are to
maximize food truck placement area and prevent ackward rendering.
If rotating the device is important and preventing awkward rendering,
then only one case exists. It effectivly limits my food truck
placement radius to the smallest screen dimension less the height
of marker.

Say the marker is p pixels tall say smallest dimension is x pixels.
Then the radius for the maximum food truck placement area
would be (x-p)/2. Rather than centering this placement circular area
it would need to be offset down by half the height of the marker.