fix: break foodTrucks paths into separate async calls
I thought a neat feature would be to use the legs of the daily
route as progress bars til the next stop. Initially the food truck's
directions were being called once and included waypoints as
intermediate stops. This was good in that the calls to google maps
were minimized; however, this method caused issues with determining
travel position of the truck on the separate legs of the daily
route.

There were many modifications to the code. Added an IIFE in the
calculateAndDisplayRoute method to store the directions for each leg
of the foodtrucks route from google map async calls. Also added a
condition to wait for all direction async calls to return before
continuing. Added a few methods to give a precise structure to the
foodTrucks routes for testing.