fix: verify properties prior to meetup creation
all meetups were added to the array. not all could be displayed on the
map and the dropdown list. not 1:1. some did not have a map marker. I
verify all properties before accepting the meetup for inclusion.
This has brought to light that some meetups exist in the same location.
need to decouple the map marker from the meetup, so that unnecessary
map markers are not added.