refactor: improve the fallback options.
On code review. App did not work well in foreign countries. Requirement
was for markers to be populated when the app starts. If there were no
meetups then the app would be useless. Async error handling is more
robust and will create random meetups in the middle of nowhere if
required.